more sanitization and create utils module
