+++ root_orchestrator/system-manager-python/ext_requests/cluster_requests.py
@@ -1,36 +1,9 @@
 import logging
-import socket
 
 import requests
 from ext_requests.apps_db import mongo_find_cluster_of_job, mongo_find_job_by_id
 from ext_requests.cluster_db import mongo_find_cluster_by_id, mongo_find_cluster_by_ip
-
-
-def is_ipv6(address):
-    """Checks if the given address is a valid IPv6 address."""
-    try:
-        socket.inet_pton(socket.AF_INET6, address)
-        return True
-    except socket.error:
-        return False
-
-
-def add_brackets_if_ipv6(address):
-    """Adds brackets to the address if it's IPv6 and doesn't have them."""
-    if is_ipv6(address) and not address.startswith("["):
-        return f"[{address}]"
-    else:
-        return address
-
-
-def is_4to6_mapped(address):
-    """Checks if the given address is 4-to-6 mapped."""
-    return is_ipv6(address) and address.startswith("::")
-
-
-def extract_v4_address_from_v6_mapped(address):
-    """Returns IPv4 address, given address is a 4-to-6 mapped IP address"""
-    return address.split(":")[3]
+from utils.network import sanitize
 
 
 def cluster_request_to_deploy(cluster_id, job_id, instance_number):
@@ -40,7 +13,7 @@ def cluster_request_to_deploy(cluster_id, job_id, instance_number):
     try:
         cluster_addr = (
             "http://"
-            + add_brackets_if_ipv6(cluster.get("ip"))
+            + sanitize(cluster.get("ip"), request=True)
             + ":"
             + str(cluster.get("port"))
             + "/api/deploy/"
@@ -60,7 +33,7 @@ def cluster_request_to_delete_job(job_id, instance_number):
     try:
         cluster_addr = (
             "http://"
-            + add_brackets_if_ipv6(cluster.get("ip"))
+            + sanitize(cluster.get("ip"), request=True)
             + ":"
             + str(cluster.get("port"))
             + "/api/delete/"
@@ -68,12 +41,13 @@ def cluster_request_to_delete_job(job_id, instance_number):
             + "/"
             + str(instance_number)
         )
+        print("Requesting:", cluster_addr)
         resp = requests.get(cluster_addr)
         print(resp)
     except Exception as e:
         logging.error(e)
         print(e)
-        print("Calling Cluster Orchestrator /api/delete not successful.")
+        print("Calling Cluster Orchestrator /api/delete job not successful.")
 
 
 def cluster_request_to_delete_job_by_ip(job_id, instance_number, ip):
@@ -81,7 +55,7 @@ def cluster_request_to_delete_job_by_ip(job_id, instance_number, ip):
         cluster = mongo_find_cluster_by_ip(ip)
         cluster_addr = (
             "http://"
-            + add_brackets_if_ipv6(cluster.get("ip"))
+            + sanitize(cluster.get("ip"), request=True)
             + ":"
             + str(cluster.get("port"))
             + "/api/delete/"
@@ -89,17 +63,18 @@ def cluster_request_to_delete_job_by_ip(job_id, instance_number, ip):
             + "/"
             + str(instance_number)
         )
+        print("Requesting:", cluster_addr)
         resp = requests.get(cluster_addr)
         print(resp)
     except Exception as e:
         logging.error(e)
-        print("Calling Cluster Orchestrator /api/delete not successful.")
+        print("Calling Cluster Orchestrator /api/delete job by ip not successful.")
 
 
 def cluster_request_to_replicate_up(cluster_obj, job_obj, int_replicas):
     cluster_addr = (
         "http://"
-        + add_brackets_if_ipv6(cluster_obj.get("ip"))
+        + sanitize(cluster_obj.get("ip"), request=True)
         + ":"
         + str(cluster_obj.get("port"))
         + "/api/replicate/"
@@ -115,7 +90,7 @@ def cluster_request_to_replicate_up(cluster_obj, job_obj, int_replicas):
 def cluster_request_to_replicate_down(cluster_obj, job_obj, int_replicas):
     cluster_addr = (
         "http://"
-        + add_brackets_if_ipv6(cluster_obj.get("ip"))
+        + sanitize(cluster_obj.get("ip"), request=True)
         + ":"
         + str(cluster_obj.get("port"))
         + "/api/replicate/"
@@ -131,7 +106,7 @@ def cluster_request_to_replicate_down(cluster_obj, job_obj, int_replicas):
 def cluster_request_to_move_within_cluster(cluster_obj, job_id, node_from, node_to):
     cluster_addr = (
         "http://"
-        + add_brackets_if_ipv6(cluster_obj.get("ip"))
+        + sanitize(cluster_obj.get("ip"), request=True)
         + ":"
         + str(cluster_obj.get("port"))
         + "/api/move/"
